pyproject.toml
[tool.poetry]
name = "telegram_boilerplate"
version = "0.1.0"
description = "Telegram boilerplate bot"
authors = ["Nikolay Pavlin <afaneor@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
django-environ = "^0.4.5"
python-telegram-bot = "^20.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
mypy = "^0.991"
isort = "^5.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
